Refactor contact page and clean up
