+++ app/(default)/page.tsx
@@ -7,7 +7,6 @@ import PageIllustration from "@/components/page-illustration";
 import Hero from "@/components/hero-home";
 import Experience from "@/components/experience";
 import Workflows from "@/components/workflows";
-import Features from "@/components/features";
 import Testimonials from "@/components/testimonials";
 import Cta from "@/components/cta";
 
@@ -18,7 +17,6 @@ export default function Home() {
       <Hero />
       <Experience />
       <Workflows />
-      <Features />
       <Testimonials />
       <Cta />
     </>
